various apps: separate computer section of diagnostic reports into battery and storage sections
- mark-scan and scan can use only the storage and not show meaningless battery info
- in the context of mark-scan and scan, what does "computer" mean anyway?
- in context of admin and central-scan, "computer" seems sort of obvious - I mean it's all on a computer!

review PDF diffs to get a sense of change
