+++ libs/ui/src/diagnostics/scan_readiness_report.test.tsx
@@ -10,10 +10,6 @@ test('ScanReadinessReport', () => {
     <ScanReadinessReport
       generatedAtTime={generatedAtTime}
       machineId={machineId}
-      batteryInfo={{
-        level: 0.5,
-        discharging: true,
-      }}
       diskSpaceSummary={{
         total: 1_000_000_000,
         available: 500_000_000,
@@ -38,11 +34,10 @@ test('ScanReadinessReport', () => {
   );
 
   screen.getByText(/Example Primary Election/);
-  screen.getByText('Battery Level: 50%');
-  screen.getByText('Power Source: Battery');
+
   screen.getByText('Free Disk Space: 50% (500 GB / 1000 GB)');
 
   screen.getByText('The printer is loaded with paper and ready to print.');
 
-  screen.getByText('Diagnostic test print successful, 1/1/2022, 12:00:00 AM');
+  screen.getByText('Test print successful, 1/1/2022, 12:00:00 AM');
 });
